pyproject.toml
[tool.poetry]
name = "blenderuserscripts"
version = "0.1.0"
description = ""
authors = ["Orito Itsuki <20170107+MatchaChoco010@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
"fake-bpy-module-3.4" = "^20230117"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
